+++ packages/heatmap-calendar/src/hooks/useDarkTheme.ts
@@ -1,0 +1,10 @@
+import { useState } from 'react';
+
+/**
+ * DarkMode 需要的状态
+ */
+export const useDarkTheme = () => {
+  const [theme] = useState<'light' | 'dark'>('light');
+
+  return { theme };
+};