+++ packages/image-gallery/src/utils/helper.ts
@@ -1,5 +1,7 @@
 /* istanbul ignore file */
 
+import { message } from 'antd';
+
 /**
  * 利用 Canvas 生成 png dataURL
  * @param image
@@ -11,7 +13,11 @@ export const getImageBase64 = (image: HTMLImageElement, scale = 1) => {
   canvas.height = image.height * scale;
 
   const context = canvas.getContext('2d');
-  context?.drawImage(image, 0, 0);
+  context?.drawImage(image, 0, 0, image.width, image.height);
 
   return canvas.toDataURL('image/png');
+};
+
+export const copySuccess = () => {
+  message.success('🎉 复制成功!');
 };