+++ packages/image-gallery/src/utils/png.ts
@@ -1,8 +1,11 @@
 /* istanbul ignore file */
 
-import { message } from 'antd';
-import { getImageBase64 } from './helper';
+import { copySuccess, getImageBase64 } from './helper';
 import { copyToClipboard } from './clipboard';
+import { svgSize } from './svg-size';
+
+// 放大 8 倍
+const SCALE = 8;
 
 /**
  * 复制 Png 到剪切板
@@ -10,22 +13,26 @@ import { copyToClipboard } from './clipboard';
  */
 export const copyPngFromSvg = async (url: string) => {
   const res = await fetch(url);
+  const size = svgSize(await res.clone().text());
+
   const svgBlob = await res.blob();
   const svgUrl = URL.createObjectURL(svgBlob);
 
-  const image = new Image();
+  const image = new Image(size.width * SCALE, size.height * SCALE);
   image.src = svgUrl;
 
   image.onload = async () => {
     const result = await fetch(getImageBase64(image));
+
     // 如果浏览器支持 navigator.clipboard 接口
     // 就使用 write 接口
     const isSuccess = await copyToClipboard('image/png', await result.blob());
+
     // 不然就用降级方案
     if (!isSuccess) {
       // 创建 image 对象
       const img = document.createElement('img');
-      img.src = getImageBase64(image, 8);
+      img.src = getImageBase64(image, SCALE);
       img.contentEditable = 'true';
       document.body.appendChild(img);
 
@@ -39,7 +46,7 @@ export const copyPngFromSvg = async (url: string) => {
       img.remove();
     }
 
-    message.success('🎉 复制成功!');
+    copySuccess();
   };
 };
 
@@ -50,16 +57,17 @@ export const copyPngFromSvg = async (url: string) => {
  */
 export const downloadPng = async (url: string, title: string) => {
   const res = await fetch(url);
+  const size = svgSize(await res.clone().text());
   const svgBlob = await res.blob();
   const svgUrl = URL.createObjectURL(svgBlob);
 
-  const image = new Image();
+  const image = new Image(size.width * SCALE, size.height * SCALE);
   image.src = svgUrl;
 
   image.onload = () => {
     const a = document.createElement('a');
     a.download = `${title}.png`;
-    a.href = getImageBase64(image, 8);
+    a.href = getImageBase64(image);
     a.click();
   };
 };