+++ packages/mindflow/src/components/Port/InPort.test.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+
+import { InPort } from './InPort';
+
+describe('InPort', () => {
+  it('默认', () => {
+    const { container } = render(<InPort color={'red'} />);
+    expect(container).toMatchSnapshot();
+  });
+});